add vuetify, create dashboard and print job pages, with routing between
