+++ vue-project/package.json
@@ -29,6 +29,6 @@
     "unplugin-vue-router": "^0.10.0",
     "vite": "^5.3.3",
     "vite-plugin-vuetify": "^2.0.3",
-    "vue-router": "^4.4.0"
+    "vue-router": "^4.4.5"
   }
 }
