+++ vue-project/src/router/index.js
@@ -7,7 +7,14 @@
 
 // Composables
 import { createRouter, createWebHistory } from 'vue-router/auto'
-import { routes } from 'vue-router/auto-routes'
+import Dashboard from '@/components/Dashboard.vue'
+import PrintJobs from '@/components/PrintJobs.vue'
+
+const routes = [
+	{ path: '/', component: Dashboard},
+	{ path: '/PrintJobs', component: PrintJobs},
+];
+
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
